Refactor: extract ModelPanel from TopBar; center sim canvas
- Extract model panel (vendor tabs, drag-and-drop reorder, custom model
  form, API key modal) into standalone ModelPanel.tsx — TopBar 663→113 lines
- Center simulation canvas in sim-stage with flex layout
- SVG canvas scales down responsively via maxWidth/maxHeight
